==> 00_Setup.ipynb <==
# Setup
> Friday August 18th, 2023

# | default_exp utils

## Previewing The Initial Datasets

> Loading Datasets using Pandas

# | export
# | hide
from nbdev.showdoc import *
import pandas as pd
import re

# | export
contact_methods = pd.read_csv('data/contact_methods.csv').fillna('')
contacts = pd.read_csv('data/contacts.csv').fillna('')
gifts = pd.read_csv('data/gifts.csv').fillna('')

contacts.head(3)

gifts.head(3)

contact_methods.head(3)

## Inital Transformations
> Transforming certain task on load in order to reduce redundency ... and some work for me :)

#### Column Name Transformation
<br>

- Nothing worse than malformed column names, amiright? <br>
<br>

# | export
def to_camel_case(s):
    """Converts a string to camel case."""
    # Remove all non-alphanumeric characters and replace with a space
    s = re.sub(r'[^a-zA-Z0-9]', ' ', s)

    # Split by space and capitalize the first letter of each word
    words = s.split()
    return ''.join(word.capitalize() for word in words)

> Examples

stripping spaces

to_camel_case("this isa test")

Stripping Special Characters

to_camel_case("this_is_a_test")

to_camel_case("hello!world_what_is_up?")

# | export
def transform_cnames(df, func=to_camel_case):
    """Apply a function to all column names of a dataframe.
    It may not always be the case that I want to apply to_camel_case as the column transformation function.
    Because of this func is an optional argument with a default value of to_camel_case  
    """
    df.columns = df.columns.map(func)
    return None

Apply a function to all column names of a dataframe

# | exports
for df in [contact_methods, contacts, gifts]:
    transform_cnames(df)

Validating that the contacts column names are now CamelCase

contacts.columns

#### Column Type Transformation

I identified 2 columns on the gift table that should be ints <br>

I'm going to convert them both to type int

# | exports
int_cols = ['GiftId', 'PledgeNumber']

# | exports
gifts[int_cols] = gifts[int_cols].replace({'': 0}).astype(int)

AmountRecieved should be a float <br>

Removing any special characters (besides dashes and periods) and converting to a float :)

# | exports
gifts['AmountReceived'] = gifts.AmountReceived.apply(
    lambda x: float(re.sub(r'[^a-zA-Z0-9\.-]', '', x)))

Previewing change

gifts.AmountReceived.head(5)

Replacing '0' values in `PledgeNumber` & `GiftId` to thier index so that each row is unique 

# | exports
gifts.loc[gifts.PledgeNumber == 0,
          'PledgeNumber'] = gifts[gifts.PledgeNumber == 0].index
gifts.loc[gifts.GiftId == 0, 'GiftId'] = gifts[gifts.GiftId == 0].index

gifts[["PledgeNumber", "GiftId"]].head(5)

Renaming the columns to match virtuous requirements

# | exports
gifts = gifts.rename(
    columns={'PledgeNumber': 'LegacyPledgeID', 'GiftId': 'LegacyGiftId'})

## Misplaced Data Transformation

> Identify and clean any pieces of data in the wrong column

contacts[["Phone", "EMail"]].head(3)

Creating a function that uses regular expressions to identify a string as a phone or a email

# | export
def classify_phone_email(value):
    """Classify a value as a phone number, email, or None."""
    if "@" in value:
        return "email"
    if re.search(r'\d{3}-\d{3}-\d{4}', value):
        return "phone"
    return None

> Examples

classify_phone_email("321-295-2502")

classify_phone_email("warren.hyson5@gmail.com")

classify_phone_email("test321-29@gmail.com")

classify_phone_email("3212952502@gmail.com")

classify_phone_email("321-295-2502@gmail.com")

#### Applying Phone/Email Transformation

Looping over phone & email records.<br>
<br>
If a phone and/or email are identified in the wrong column, I'm swapping thier values

# | export
for index, row in contacts.iterrows():
    phone_classification = classify_phone_email(row['Phone'])
    email_classification = classify_phone_email(row['EMail'])

    if phone_classification == "email":
        contacts.at[index, 'EMail'] = row['Phone']
        contacts.at[index, 'Phone'] = ''

    if email_classification == "phone":
        contacts.at[index, 'Phone'] = row['EMail']
        contacts.at[index, 'EMail'] = ''

contacts[["Phone", "EMail"]].head(3)

## Consolidate Contacts Table

- Searching for missing users
- Splitting / joining households


#### Checking for Missing records

Checking for any DonorNumbers that are NOT IN contacts Number...

(~gifts.DonorNumber.isin(contacts.Number.unique())).any()

Found 1! <br>
<br>
Doing the same for contact_methods

(~contact_methods.DonorNumber.isin(contacts.Number.unique())).any()

Didn't find any there <br>
Going to extract the missing records from the gifts table

# | exports
donors_not_in_contacts = gifts.loc[~gifts.DonorNumber.isin(
    contacts.Number.unique()), :]

donors_not_in_contacts

Adding Shakespeare to our contacts!

# | exports
contacts = pd.concat([
    contacts,
    # Dataframe of donors not in contacts
    pd.DataFrame(donors_not_in_contacts[['DonorNumber', 'FirstName', 'LastName']]
                 .drop_duplicates()
                 .rename(columns={'DonorNumber': 'Number'})
                 .drop_duplicates()
                 .to_dict('records'))
])

#### Spliting rows with Multiple People

contacts.FirstName.head(1)

<br>Split the names on ' & ' or ' and ', then expand the resulting lists into new rows

# | exports
contacts[['FirstName', 'SecondaryFirstName']] = contacts['FirstName'].str.split(
    ' & | and ', expand=True).fillna('')

Aditionally, going to look for any records where there's a duplicated `Number` <br>

The following Python code will return `True` for rows where the `Number` was deemed to be a duplicate:
 
```Python
contacts.Number.duplicated()
```

For all the rows that are duplicates, I'm going to:<br><br>
1. Place them in a seperate variable <br>
2. Remove them from the original dataset <br>
3. Add them as the `2nd Person` on the non-duplicated row <br>
4. Ensure they have proper first & last names

# | exports
records_to_join = contacts.loc[contacts.Number.duplicated(), :].to_dict(
    orient='records')

 Removing duplicates from contacts before joining

# | exports
contacts = contacts.loc[~contacts.Number.duplicated(), :]

Adding dupliates as secondary contacts

# | exports
for record in records_to_join:
    contacts.loc[contacts.Number.isin([record['Number']]), [
        'SecondaryFirstName', 'SecondaryLastName']] = [record['FirstName'], record['LastName']]

# | export
contacts['SecondaryLastName'] = contacts.SecondaryLastName.fillna('')

Adding Secondary Last Name where it was previously missing

# | exports
contacts['SecondaryLastName'] = contacts.apply(
    lambda x: x['LastName'] if x['SecondaryLastName'] == '' and x['SecondaryFirstName'] != '' else x['SecondaryLastName'], axis=1)

#### Updating unique identifiers
> Ensuring LegacyIndividualId is present for all contacts & is unique

Initalizing empty collumns

# | export
contacts[['LegacyIndividualId', 'SecondaryLegacyIndividualId']] = None

# | export
contacts.reset_index(inplace=True, drop=True)

Adding id since none was provided

# | exports
id = 0
for index, row in contacts.iterrows():
    contacts.loc[index, 'LegacyIndividualId'] = id
    id += 1
    if row['SecondaryFirstName'] != '':
        contacts.loc[index, 'SecondaryLegacyIndividualId'] = id
        id += 1

# | export
contacts.fillna('', inplace=True)

## Cleaning Contact First & Last Names
>Cleaning Records with blank first or last names <br>

Checking for records where FirstName  and/or LastName is blank


# | exports
blank_name_records = ((contacts.FirstName == '') | (contacts.LastName == ''))

Previewing the blank name records

contacts.loc[((contacts.FirstName == "") | (contacts.LastName == "")), :]

Before I delete the records I'm going to check if the names are present on the gift table <br>
<br>
I'm going to start by getting the unique Numbers that the records belong too

# | exports
blank_name_numbers = contacts.loc[blank_name_records, 'Number']

gifts.loc[gifts.DonorNumber.isin(blank_name_numbers), [
    'DonorNumber', 'FirstName', 'LastName']].drop_duplicates().head(5)

The names are present on the gifts table!

# | exports
gift_name_records = gifts.loc[gifts.DonorNumber.isin(
    blank_name_numbers), ['DonorNumber', 'FirstName', 'LastName']].drop_duplicates()

Removing the invalid record

# | exports
gift_name_records = gift_name_records.loc[(
    (gift_name_records.FirstName != '') & (gift_name_records.LastName != '')), :]
gift_name_records

Updating the records that previousuly had a blank first or last name

# | exports
for _, row in gift_name_records.iterrows():
    contacts.loc[contacts['Number'] == row['DonorNumber'], [
        'FirstName', 'LastName']] = [row['FirstName'], row['LastName']]

All the records valid names now!

contacts.loc[blank_name_records, :]

## Add ContactName Column

contacts.head(3)

Creating a function to update the contact name depending on the individuals on the household record

# | exports
def set_contact_name(row):
    """Sets the contact name for a given row in the dataframe."""
    if row['LastName'] == row['SecondaryLastName']:
        return row['FirstName'] + ' & ' + row['SecondaryFirstName'] + ' ' + row['LastName']
    elif row['SecondaryFirstName'] != '':
        return row['FirstName'] + ' ' + row['LastName'] + ' & ' + row['SecondaryFirstName'] + ' ' + row['SecondaryLastName']
    else:
        return row['FirstName'] + ' ' + row['LastName']

# | exports
contacts['ContactName'] = contacts.apply(set_contact_name, axis=1)

Previewing result

contacts[['ContactName', 'FirstName', 'LastName',
          'SecondaryFirstName', 'SecondaryLastName']].head(5)

## ProjectCode Transformation
>Splitting FundId

Splitting the FundId column by `','`

# | exports
project_codes = gifts.FundId.str.split(', ', expand=True)

project_codes.head(6)

Assigning the values to the Project1 & 2 columns

# | exports
gifts[['Project1Code', 'Project2Code']] = project_codes

Removing FundId from the dataset

# | exports
gifts = gifts.loc[:, gifts.columns.drop('FundId')].copy()

## Date Transformation
>Formating date

df["Date"].head(5)

<br>Since Dates are in different formats I'm going to create a custom parser

# | export
from datetime import datetime

# | export
def custom_parser(date_str):
    """Parses a date string into a datetime object."""
    try:
        return datetime.strptime(date_str, '%m/%d/%Y')
    except ValueError:
        return datetime.strptime(date_str, '%Y/%m/%d')

> Examples

custom_parser("4/14/1999")

custom_parser("1999/04/14")

#### Transforming Dates

# | exports
gifts['GiftDate'] = gifts['Date'].apply(custom_parser)

Dropping old date columns

# | exports
gifts = gifts.loc[:, gifts.columns.drop('Date')].copy()

## Email Transformation
>Validate emails

Previewing the `Email` column

contacts.EMail.head(1)

<br>It looks like the first email 'kklumbers@ yahoo.co' is invalid and should be fixed <br>
<br>
I'm going to write a function to validate emails and fix common mistakes

# | export
def valid_email(s):
    """Validates an email address."""
    pattern = re.compile(r'^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$')
    if pattern.match(s):
        return True
    return False

### Examples

valid_email("warren.hyson5@gmail.com")

valid_email("warren.hyson5@gmail")

valid_email("warren.hyson5@.com")

valid_email("warren.hyson5@gmailcom")

valid_email("warren.hyson5@@gmail.com")

<br> Creating a second function that attempts to fix common email mistakes like mispelled domains

# | export
def fix_email(email):
    """Fixes common mistakes in email addresses."""
    if not email:
        return ''

    email = re.sub(r'\s', '', email)

    # Check and fix common mistakes in top-level domain (TLD)
    for wrong_tld, correct_tld in [('co', 'com'), ('cmo', 'com'), ('con', 'com')]:
        if email.endswith('.' + wrong_tld):
            email = email[:-len(wrong_tld)] + correct_tld

    # Check and fix common mistakes in domain names
    for wrong_domain, correct_domain in [('gmial', 'gmail'), ('yahho', 'yahoo')]:
        email = email.replace('@' + wrong_domain, '@' + correct_domain)

    if valid_email(email):
        return email

    return ''

### Examples

`correct`

fix_email("warren.hyson5@gmail.com")

`gmail` spelled wrong

fix_email("warren.hyson5@gmial.com")

`com` spelled wrong

fix_email("warren.hyson5@gmial.con")

#### Applying Email Transformation

Previewing

contacts.EMail.head(1).apply(fix_email)

# | exports
contacts['EMail'] = contacts.EMail.apply(fix_email)

## Phone # Transformation
>Fix phone numbers

There appears to be a invalid phone number

contacts.loc[2, "Phone"]

Writing a function to validate US phone numbers, assuming the phone numbers provided are from the US

# | export
def validate_us_phone_number(phone_number):
    """Validates a US phone number and returns it if valid, otherwise returns an empty string."""
    # Patterns for different US phone number formats
    patterns = [
        r'^\+1\s?\d{3}-\d{3}-\d{4}$',
        r'^\(\d{3}\)\s?\d{3}-\d{4}$',
        r'^\d{3}-\d{3}-\d{4}$',
        r'^\d{3}-\d{4}$'
    ]

    # Check if the phone number matches any of the patterns
    for pattern in patterns:
        if re.match(pattern, phone_number):
            return phone_number

    # Checking if the string contains only numbers and is of length 7 or 10
    numeric_string = re.sub('[^0-9]', '', phone_number)
    length = len(numeric_string)
    if length == 7 or length == 10:
        return phone_number

    return ''

> Examples

validate_us_phone_number("321-295-2502")

`no-dashes` 10 digit

validate_us_phone_number("3212952502")

`no-dashes` 7 digit

validate_us_phone_number("2952502")

`no-dashes` 11 digit

validate_us_phone_number("13212952502")

`With Parenthesese`

validate_us_phone_number("(123) 456-7890")

`Random Text`

validate_us_phone_number("407-555-1234 ext. 1234")

#### Applying Phone # Transformation

# | exports
contacts['Phone'] = contacts.Phone.apply(validate_us_phone_number)

# | exports
contacts.fillna('', inplace=True)

# Results

contacts.head(5)

gifts.head(5)

contacts.head(5)

# | hide
import nbdev

# | hide
nbdev.nbdev_export("00_Setup.ipynb")

<br>
