==> 02_Pandas_Solution.ipynb <==
# Python Solution
> Performing the data cleaning / ETL using Pandas <br>
<br>

- I'm going to demonstrate how pandas can be used the finish the rest of the ETL process

#| default_exp solution_pd

#| export
# | hide
import pandas as pd
import numpy as np
import re

from virtuous_interview.utils import contacts, contact_methods, gifts
from nbdev.showdoc import *

# Solution

> `Pandas :)`

## Private

> Does someone want to be private


Source Table:

- Contacts Table

Solution:

1. Create procedure to add new column Private

Looking for notes

contacts.Remarks[~contacts.Remarks.isin([''])]

contacts['Private'] = contacts.Remarks.apply(lambda x: True if x == 'Is anonymous' else False)

## ContactType

> is required and can only be Household or Organization

Source Table:

- Contacts Table

Solution:

1. Create procedure to add new column ContactType

contacts[['Number', 'CompanyName']]

<br>

#| exports
contacts['ContactType'] = contacts.apply(lambda x: 'Household' if x['CompanyName'] == '' else 'Organization', axis=1)

contacts[['Number', 'CompanyName', 'ContactType']].head(3)

## Postal Code

> if address is present and is US, must be a valid zip code, either 12345 or 12345-1234


Source Table:

- Contacts

Solution:

1. Create procedure to remove any postal codees that doesn't match the approved format from the [usps](https://pe.usps.com/archive/html/dmmarchive20030810/A010.htm)


contacts['Postal'].head(3)

#| exports
postal_code_pattern = '[0-9]{5}(?:-[0-9]{4})?$'

### Examples

bool(re.match(postal_code_pattern, '32828'))

bool(re.match(postal_code_pattern, '328289'))

bool(re.match(postal_code_pattern, '3282'))

bool(re.match(postal_code_pattern, '32828-237'))

bool(re.match(postal_code_pattern, '32828-2379'))

### Applying Transformation

#| exports
contacts['Postal'] = contacts.Postal.apply(lambda x: x if re.match(postal_code_pattern, x) else '')


contacts['Postal'].head(3)

## IsDeceased

> can only be TRUE or FALSE

Source Table: 

- Contacts

Solution:

1. Create procedure to update Deceased to TRUE/FALSE


contacts.Deceased.head(3)

#| exports
contacts['Deceased'] = contacts.Deceased.apply(lambda x: True if x == 'Yes' else False)


contacts.Deceased.head(3)

## GiftType

> Can only be Cash, Check, Credit, Other, or Reversing Transaction

Source Table: 

- Gifts

Solution:

1. Identify Incorrect Gift Types
2. Create procedure to replace invalid gift types


gifts[['AmountReceived', 'PaymentMethod']].tail(10)

#| exports
def clean_payment_type(row):
    payment_method = ''
    orginal_payment_method = str(row['PaymentMethod']).lower()
    
    if row['AmountReceived'] < 0:
        payment_method = 'Reversing Transaction'
    elif re.match('credit', orginal_payment_method):
        payment_method = 'Credit'
    elif orginal_payment_method in ['check', 'cash', 'reversing transaction']:
        payment_method = orginal_payment_method.title()
    else:
        payment_method = 'Other'

    return payment_method

## Applying GiftType Transformation

#| exports
gifts['PaymentMethod'] = gifts.apply(clean_payment_type, axis=1)


gifts.PaymentMethod.tail(10)

## CreditCardType

> Can only be Visa, Mastercard, AMEX, Discover

Solution:

1. Identify Incorrect Credit Types
2. Create procedure to replace invalid credit types


gifts.CreditCardType.unique()

Creating functions to validate each of the credit card types

#| export
def validate_mastercard(string):
    # Match strings that contain "mastercard" or "master card", case insensitive
    pattern = r'master\s?card'
    return bool(re.search(pattern, string, re.IGNORECASE))

#| export
def validate_amex(string):
    # Match strings that contain "amex", "american express", or "american ex", case insensitive
    pattern = r'amex|american\s?express|american\s?ex'
    return bool(re.search(pattern, string, re.IGNORECASE))

#| export
def validate_visa(string):
    # Match strings that contain "visa", case insensitive
    pattern = r'visa'
    return bool(re.search(pattern, string, re.IGNORECASE))

#| export
def validate_discover(string):
    # Match strings that contain "discover", case insensitive
    pattern = r'discover'
    return bool(re.search(pattern, string, re.IGNORECASE))

#| export
def validate_credit_card(string): 
    if validate_mastercard(string):
        return 'Mastercard'
    elif validate_amex(string):
        return 'AMEX'
    elif validate_visa(string):
        return 'Visa'
    elif validate_discover(string):
        return 'Discover'
    else:
        return ''

validate_credit_card('Amex')

validate_credit_card('Master Card')

validate_credit_card('visa')

validate_credit_card('test')

validate_credit_card('Gold card')

## Applying CreditCardType Transformation

#| exports
gifts['CreditCardType'] = gifts.CreditCardType.apply(validate_credit_card)

# Creating Final Datasets

## Contact Methods

> Table of constituent contact information

Merging Contact and Contact Methods Tables To get ALL the contact information

#| exports
temp_contacts = contacts[['Number', 'Phone', 'EMail']].rename(columns={
    'Number': 'LegacyContactId',
    'Phone': 'HomePhone',
    'EMail': 'HomeEmail'
}).copy()

#| exports
temp_contact_methods = contact_methods.rename(columns={'DonorNumber':'LegacyContactId'})

#| exports
contacts_wide = temp_contacts.merge(temp_contact_methods,
                                how='left',
                                on='LegacyContactId',
                                ).sort_values('LegacyContactId')

Pivoting the Data, converting the HomePhone, HomeEmail, And Fax values to individual Type : Value pairs

contacts_v = contacts_wide.melt(id_vars=['LegacyContactId'], value_vars=['HomePhone', 'HomeEmail', 'Fax', 'Phone', 'EMail'], var_name='Type', value_name='Value')

Creating a function to safley check for nan values

def is_nan(x):
    try:
        return np.isnan(x)
    except TypeError:
        return False

final_contact_methods = contacts_v[~((contacts_v.Value == '') | (contacts_v.Value.apply(is_nan)))].reset_index(drop=True).copy()

def clean_contact_type(s):
    if s == 'Phone':
        return 'HomePhone'
    elif s == 'EMail':
        return 'HomeEmail'
    else:
        return s

final_contact_methods['Type'] = final_contact_methods.Type.apply(clean_contact_type)

final_contact_methods.drop_duplicates(inplace=True)

final_contact_methods.sort_values('LegacyContactId').reset_index(drop=True)

<br>

final_contact_methods.to_csv('data/final_contact_methods.csv', index=False)

## Contacts

> Table of constituent contact information

Checking for any missing contacts

def is_empty(x):
    if str(x) == '' or str(x) == 'nan':
        return True
    else:
        return False

def format_string(x):
    if is_empty(x):
        return ''
    else: 
        return x

contacts_wide['HomePhone'] = np.where(contacts_wide['HomePhone'].apply(is_empty), format_string(contacts_wide['Phone']), format_string(contacts_wide['HomePhone']))
contacts_wide['HomeEmail'] = np.where(contacts_wide['HomeEmail'].apply(is_empty), format_string(contacts_wide['EMail']), format_string(contacts_wide['HomeEmail']))


contacts_wide = contacts_wide[['LegacyContactId', 'HomePhone', 'HomeEmail']].fillna('').copy()

contacts_wide = contacts_wide[~((contacts_wide.HomePhone == '') & (contacts_wide.HomeEmail == ''))].copy()

contacts = (contacts
            .drop(['Phone', 'EMail'], axis=1)
            .merge(contacts_wide,
                   left_on='Number',
                   right_on='LegacyContactId',
                   how='left',
            )
            .drop('LegacyContactId', axis=1)
            )

columns = [
    'LegacyContactId', 'LegacyIndividualId', 'ContactType', 'ContactName',
    'FirstName', 
    'LastName', 'SecondaryLegacyIndividualId', 'SecondaryFirstName',
    'SecondaryLastName', 'HomePhone', 'HomeEmail', 'Address1', 
    'City', 'State', 'PostalCode', 'IsPrivate', 'IsDeceased',
    ]

contacts.rename(columns={
    'Number':'LegacyContactId',
    'Street': 'Address1',
    'Postal': 'PostalCode',
    'Private': 'IsPrivate',
    'Deceased': 'IsDeceased',
}, inplace=True)

final_contacts = contacts[columns].fillna('')

final_contacts

<br>

final_contacts.to_csv('data/final_contacts.csv', index=False)

## Gifts

> Table with gift history

columns = ['LegacyContactId', 'LegacyGiftId', 'GiftType', 'GiftDate',
           'GiftAmount', 'Notes', 'CreditCardType', 'Project1Code',
           'Project2Code', 'LegacyPledgeID']

gifts.rename(columns={
    'DonorNumber': 'LegacyContactId',
    'PaymentMethod': 'GiftType',
    'AmountReceived': 'GiftAmount',
}, inplace=True)

gifts[columns]

<br>

final_gifts = gifts[columns].copy()

final_gifts.to_csv('data/final_gifts.csv', index=False)

# | hide
import nbdev

# | hide
nbdev.nbdev_export('02_Pandas_Solution.ipynb')
