==> 02_SQL_Solution.ipynb <==
#| default_exp solution_sql

# SQL Solution
`Creating Series of SQL Procedures to use for the Data Migration` <br>
<br>
- I'm going to use a local MySQL instance as a DEV environment <br>
- Create a new Database for the purposes of this project <br>
- Load the inital datasets into MySQL with the 'temp_' prefix <br>
- Go through the modifications highlighted on the Requirements <br>
    - For each modification I will write a stored procedure that I use @ the end <br>
    - For some modifications I will write a view to demonstrate the logic is correct before proceeding <br>

#| export
import pandas as pd
import mysql.connector

from virtuous_interview.utils import contacts, contact_methods, gifts
from mysql.connector import Error
from dotenv import dotenv_values
from sqlalchemy import create_engine

# Configuring SQL
`Creating a new db instance and loading tables`

## Creating DB
`exam_db`

#| export
DB_NAME = 'exam_db'

Using `mysql` library to connect as root user and create a new db instance

#| export solution_sql
try:
    connection = mysql.connector.connect(user='root', host='localhost')
    cursor = connection.cursor()
    cursor.execute(f"CREATE DATABASE IF NOT EXISTS {DB_NAME};")
except Error as e:
    print(f"Error: {e}")
finally:
    if connection.is_connected():
        cursor.close()
        connection.close()


## execute_sql
`Function to execute sql on the DEV db`

Since that was a lot of code just to execute 1 line of SQL, I'm going to create a new function that will make it easier

#| export
DB_USER = dotenv_values()['DB_USER']
DB_PASSWORD = dotenv_values()['DB_PASSWORD']

#|export solution_sql
def insert_sql(sql):
    try:
        # Connect to MySQL
        connection = mysql.connector.connect(user=DB_USER, password=DB_PASSWORD, host='localhost', database=DB_NAME)
        cursor = connection.cursor()
    
        # Execute the SQL command to create the view
        cursor.execute(sql)
        print("SQL executed successfully")
    
    except Error as e:
        print(f"Error: {e}")
    
    finally:
        if connection.is_connected():
            cursor.close()
            connection.close()
        print("MySQL connection closed")

## insert_proc
`Function to insert procedure into the Database` <br>
<br>
- Useful since I'll need to create many procedures

#|export solution_sql
def insert_proc(sql, proc_name, call=True):
    insert_sql(f'DROP PROCEDURE IF EXISTS {proc_name};')
    stmt = f"""         
        CREATE PROCEDURE {proc_name}()
        BEGIN
            {sql}
        END;
        COMMIT;
    """
    insert_sql(stmt)
    if call:
        insert_sql(f'CALL {proc_name}();')

## Loading Datasets Into Database
`Using prefix 'temp_'`

Mapping DataFrames to table names

#| export solution_sql
tables = {'temp_contact_methods': contact_methods, 'temp_contacts': contacts, 'temp_gifts': gifts}

#| export
engine = create_engine(f'mysql+mysqlconnector://{DB_USER}:{DB_PASSWORD}@localhost/{DB_NAME}')

#| export solution_sql
try:
    for table_name, df in tables.items():
        try:
            df.to_sql(table_name, engine, index=False, if_exists='fail')
        except ValueError:
            print(f"Table {table_name} already exists. Skipping.")

except Error as e:
    print(f"Error: {e}")

finally:
    if connection.is_connected():
        cursor.close()
        connection.close()


Previewing the data

#| hide
pd.read_sql('select * from temp_gifts limit 5', engine)

# Solution
`Creating procedures for each modification`

## ContactType
`is required and can only be Household or Organization`<br>
<br>
- Source Table: Contacts Table
- Solution:
    - Create procedure to add new column ContactType

#| hide
contacts_view = f"""
    CREATE or REPLACE VIEW contacts_view AS
    SELECT Number, CompanyName, CASE WHEN CompanyName = '' THEN 'Household' ELSE 'Organization' END AS ContactType
    FROM temp_contacts
"""
insert_sql(contacts_view)
pd.read_sql('select * from contacts_view limit 5', con=engine)


Now I'm going to write a procedure that would perform this transformation on the temp table. <br>
<br>
Since it's possible that I may need to run this procedure multiple times, each time the data is updated I'm going to write **2** procedures to solve this problem <br>
<br>
1. Add column procedure
    - Add a column if it doesn't exist
2. Procedure to add ContactType

#| export solution_sql
add_column = """
    DROP PROCEDURE IF EXISTS add_column;
    
    
    CREATE PROCEDURE add_column(
        IN tableName VARCHAR(255),
        IN columnName VARCHAR(255),
        IN columnType VARCHAR(255)
    )
    BEGIN
        DECLARE columnExists BOOLEAN DEFAULT FALSE;
    
        SELECT COUNT(*)
        INTO columnExists
        FROM information_schema.COLUMNS
        WHERE TABLE_SCHEMA = DATABASE() AND TABLE_NAME = tableName AND COLUMN_NAME = columnName;
    
        IF columnExists = 0 THEN
            SET @sql = CONCAT('ALTER TABLE ', tableName, ' ADD COLUMN ', columnName, ' ', columnType);
            PREPARE stmt FROM @sql;
            EXECUTE stmt;
            DEALLOCATE PREPARE stmt;
        END IF;
        COMMIT;
    END;
"""
insert_sql(add_column)

#|export solution_sql
insert_contact_type = """
    CALL add_column('temp_contacts', 'ContactType', 'VARCHAR(255)');
    
    UPDATE temp_contacts
    SET ContactType = CASE WHEN CompanyName = '' THEN 'Household' ELSE 'Organization' END;
    COMMIT;
"""
insert_proc(insert_contact_type, 'insert_contact_type', call=True)

## Private
`Does someone want to be private`<br>
<br>
- Source Table: Contacts Table
- Solution:
    - Create procedure to add new column Private

insert_private = """
    CALL add_column('temp_contacts', 'Private', 'TINYINT');

    UPDATE temp_contacts
    SET Private = CASE WHEN Remarks = 'Is anonymous' THEN 1 ELSE 0 END;
    commit;
"""
insert_proc(insert_private, 'insert_private', call=True)

## Postal Code
> if address is present and is US, must be a valid zip code, either 12345 or 12345-1234 <br>
<br>
- Source Table: Contacts
- Solution:
    - Create procedure to remove any postal codees that doesn't match the approved format from the [usps](https://pe.usps.com/archive/html/dmmarchive20030810/A010.htm)

#| hide
','.join(contacts.columns.drop('Postal').tolist())

#| hide
contacts_view = f"""
    CREATE OR REPLACE VIEW contacts_view AS
    SELECT
        NUMBER,
        POSTAL AS OLD_POSTAL,
        CASE
        WHEN REGEXP_LIKE(Postal, '^[0-9]{{5}}$') OR REGEXP_LIKE(Postal, '^[0-9]{{5}}-[0-9]{{4}}$') THEN Postal
        ELSE ''
        END AS NEW_POSTAL
    FROM temp_contacts;
"""
insert_sql(contacts_view)
pd.read_sql("SELECT * FROM contacts_view where old_postal != '' limit 5", con=engine)

Success!<br>
<br>
Creating Stored Procedure...

#|export solution_sql
update_zip = """
    UPDATE temp_contacts
    SET Postal = ''
    WHERE Postal NOT REGEXP '^[0-9]{5}$' AND Postal NOT REGEXP '^[0-9]{5}-[0-9]{4}$';
    COMMIT;
"""
insert_proc(update_zip, 'update_zip', call=True)

## IsDeceased
`can only be TRUE or FALSE`<br>
<br>
- Source Table: Contacts <br>
- Solution: <br>
    - Create procedure to update Deceased to TRUE/FALSE

#|export solution_sql
update_deceased = f"""
    UPDATE temp_contacts
    SET Deceased = CASE
        WHEN Deceased = 'Yes' THEN 1
        ELSE 0
    END;
    commit;
    ALTER TABLE temp_contacts MODIFY Deceased TINYINT;
    commit;
"""
insert_proc(update_deceased, 'update_deceased', call=True)

## GiftType
`Can only be Cash, Check, Credit, Other, or Reversing Transaction`<br>
<br>
- Source Table: Gifts
- Solution:
    - Identify Incorrect Gift Types
    - Create procedure to replace invalid gift types

#| hide
pd.read_sql('Select Distinct PaymentMethod from temp_gifts', engine)

It looks like there are several payment methods that don't match the approved list. Additionally, the payment method 'credit card' will need to be mapped to 'credit'

#|export solution_sql
update_gift_type = f"""
  UPDATE temp_gifts
  SET PaymentMethod = CASE
    WHEN AmountReceived < 0 THEN 'Reversing Transaction'
    WHEN LOWER(TRIM(PaymentMethod)) = 'cash' THEN 'Cash'
    WHEN LOWER(TRIM(PaymentMethod)) = 'check' THEN 'Check'
    WHEN LOWER(TRIM(PaymentMethod)) LIKE 'credit%' THEN 'Credit'
    ELSE 'Other'
  END;
  commit;
"""
insert_proc(update_gift_type, 'update_gift_type', call=True)

## CreditCardType
`Can only be Visa, Mastercard, AMEX, Discover` <br>
<br>
- Solution: <br>
    - Identify Incorrect Credit Types <br>
    - Create procedure to replace invalid credit types

#| hide
pd.read_sql('select distinct CreditCardType from temp_gifts', engine)

#|export solution_sql
proc = f"""
UPDATE temp_gifts
    SET CreditCardType = CASE
    WHEN CreditCardType  = 'Master card' THEN 'Mastercard'
    else 'AMEX'
    end
WHERE CreditCardType IN ('American Ex', 'Master car');
commit;
"""
insert_proc(proc, 'update_gift_type', call=True)

# Execution
`Creating Final Tables` <br>
<br>

[contact_methods, contacts, gifts]

## Contacts
`Creating the final contacts table` <br>
<br>

- The cleaning is done <br>
- Creating contacts table and renaming various columns <br>

pd.read_sql('SELECT * FROM temp_contacts', engine)

pd.read_sql("""
            SELECT 
                `Number`,
                `LegacyIndividualId`,
                `ContactType`
                `ContactName`,
                `FirstName`,
                `LastName`
                `SecondaryLegacyIndividualId`,
                `SecondaryFirstName`,
                `SecondaryLastName`,
                CASE WHEN
                    temp_contact_methods.`Phone`  = '' 
                    THEN temp_contacts.`Phone`
                    ELSE temp_contact_methods.`Phone`
                END AS HomePhone,
                CASE WHEN
                    temp_contact_methods.`EMail`  = ''
                    THEN temp_contacts.`EMail`
                    ELSE temp_contact_methods.`EMail`
                END AS HomeEmail,
                `Street`,
                `City`,
                `State`,
                `Postal`,
                `Private`,
                `Deceased`
            FROM 
            temp_contacts
            LEFT JOIN
                temp_contact_methods ON temp_contact_methods.DonorNumber = temp_contacts.`Number`;
            """,
            engine
            )

proc = """


CREATE TABLE IF NOT EXISTS contacts (
    `LegacyContactId` VARCHAR(255),
    `LegacyIndividualId` VARCHAR(255),
    `ContactType` VARCHAR(255),
    `ContactName` VARCHAR(255),
    `FirstName` VARCHAR(255),
    `LastName` VARCHAR(255),
    `SecondaryLegacyIndividualId` VARCHAR(255),
    `SecondaryFirstName` VARCHAR(255),
    `SecondaryLastName` VARCHAR(255),
    `HomePhone` VARCHAR(255),
    `HomeEmail` VARCHAR(255),
    `Address1` VARCHAR(255),
    `City` VARCHAR(255),
    `State` VARCHAR(255),
    `PostalCode` VARCHAR(255),
    `IsPrivate` VARCHAR(255),
    `IsDeceased` VARCHAR(255)
);
insert into contacts
SELECT 
    `Number` AS `LegacyContactId`,
    `LegacyIndividualId`,
    `ContactType`,
    `ContactName`,
    `FirstName`,
    `LastName`,
    `SecondaryLegacyIndividualId`,
    `SecondaryFirstName`,
    `SecondaryLastName`,
    CASE WHEN
        temp_contact_methods.`Phone`  = '' 
        THEN temp_contacts.`Phone`
        ELSE temp_contact_methods.`Phone`
    END AS HomePhone,
    CASE WHEN
        temp_contact_methods.`EMail`  = ''
        THEN temp_contacts.`EMail`
        ELSE temp_contact_methods.`EMail`
    END AS HomeEmail,
    `Street` AS `Address1`,
    `City`,
    `State`,
    `Postal` AS `PostalCode`,
    `Private` AS `IsPrivate`,
    `Deceased` AS `IsDeceased`
FROM 
temp_contacts
LEFT JOIN
    temp_contact_methods ON temp_contact_methods.DonorNumber = temp_contacts.`Number`;
    
commit;
"""
insert_proc(proc, 'create_contacts', call=True)

pd.read_sql('select * from contacts limit 5', engine)

## Gifts
`Creating the final gifts column`<br>
<br>

- The cleaning is done <br>
- Creating the gifts table and renaming some columns <br>

proc = """
CREATE TABLE IF NOT EXISTS gifts (
    LegacyContactId VARCHAR(255),
    LegacyGiftId INTEGER,
    GiftType TEXT,
    GiftDate TEXT,
    GiftAmount REAL,
    Notes TEXT,
    CreditCardType TEXT,
    Project1Code TEXT,
    Project2Code TEXT,
    LegacyPledgeID INTEGER
);

insert into gifts
select 
    DonorNumber as LegacyContactId,
    LegacyGiftId,
    PaymentMethod as GiftType,
    GiftDate,
    AmountReceived as GiftAmount,
    Notes,
    CreditCardType,
    Project1Code,
    Project2Code,
    LegacyPledgeID
from temp_gifts;

commit;
"""
insert_proc(proc, 'create_gifts', call=True)

pd.read_sql('select * from gifts limit 5', engine)

## Contact Methods

#| hide
pd.read_sql('select * from contacts limit 5', engine)

#| hide
pd.read_sql('select * from temp_contact_methods limit 5', engine)

`Procedure to create contact_methods table`<br>
<br>

- Creating a contact_methods <br>
- Inserting distinct values from the contacts table <br>

pd.read_sql("""
    SELECT DISTINCT
        contacts.`LegacyContactId` AS LegacyContactId,
        contacts.`HomePhone` AS HomePhone,
        contacts.`HomeEmail` AS HomeEmail,
        temp_contact_methods.Fax AS fax
    FROM 
        contacts
    JOIN
        temp_contact_methods ON temp_contact_methods.DonorNumber = contacts.`LegacyContactId`;
            """, engine)

#|export solution_sql
proc = """
CREATE TABLE IF NOT EXISTS contact_methods (
    `LegacyContactId` VARCHAR(255),
    `Type` VARCHAR(255),
    `Value` VARCHAR(255)
);

BEGIN
DECLARE done INT DEFAULT FALSE;
DECLARE v_LegacyContactId VARCHAR(255);
DECLARE v_HomePhone VARCHAR(255);
DECLARE v_HomeEmail VARCHAR(255);
DECLARE v_Fax VARCHAR(255);
DECLARE cur CURSOR FOR 
    SELECT DISTINCT
        contacts.`LegacyContactId` AS LegacyContactId,
        contacts.`HomePhone` AS HomePhone,
        contacts.`HomeEmail` AS HomeEmail,
        temp_contact_methods.Fax AS fax
    FROM 
        contacts
    JOIN
        temp_contact_methods ON temp_contact_methods.DonorNumber = contacts.`LegacyContactId`;

DECLARE CONTINUE HANDLER FOR NOT FOUND SET done = 1;

    OPEN cur;

    read_loop: LOOP
        FETCH cur INTO v_LegacyContactId, v_HomePhone, v_HomeEmail,v_Fax;
        
        IF done THEN
            LEAVE read_loop;
        END IF;

        INSERT INTO contact_methods (`LegacyContactId`, `Type`, `Value`) VALUES 
            (v_LegacyContactId, 'HomePhone', v_HomePhone),
            (v_LegacyContactId, 'HomeEmail', v_HomeEmail),
            (v_LegacyContactId, 'Fax', v_Fax);
    END LOOP;

    CLOSE cur;
END;

-- Delete records from contact methods where value is null or ''
DELETE FROM contact_methods WHERE `Value` IS NULL OR `Value` = '';


commit;


"""
insert_proc(proc, 'transform_contact_methods', call=True)

pd.read_sql('select * from contact_methods order by LegacyContactId', con=engine)

# Export

#| hide
import nbdev

#| hide
nbdev.nbdev_export('02_SQL_Solution.ipynb')
